requirements.txt
Django==2.0.5
djangorestframework==3.9.0
SpeechRecognition==3.6.3

